backend/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
psycopg2-binary>=2.9.0
requests>=2.25.0
Pillow>=8.0.0
python-dotenv>=0.19.0 